* Hopefully fixed compilation problem
git-svn-id: svn://svn.code.sf.net/p/gmerlin/code/trunk@1397 214d8ac5-8513-4f59-80f8-2f39b674a24a

+++ lemuria/lib/xaos/src/include/i386/ctrl87.h
@@ -100,8 +100,8 @@
    -------------
   The bits in the value returned reflect the new floating-point control word.
 */
-EXTERN unsigned short _control87 (unsigned short, unsigned short);
-
+EXTERN unsigned int _control87 (unsigned int, unsigned int);
+// EXTERN unsigned short _control87 (unsigned short, unsigned short);
 
 /**************************************************************************/
 /**************************   Global variables   **************************/
